pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wagtail-spa-integration"
version = "3.3.0"
description = "Tools for using Wagtail API with JavaScript single page apps"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    'Environment :: Web Environment',
    'Framework :: Django',
    'Framework :: Wagtail',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Intended Audience :: Developers',
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "wagtail (>=6.3)",
    "Django (>=4.2)",
    "wagtail-headless-preview (>=0.8,<0.9)",
    "django-filter (>=25.1,<26)",
]
requires-python = ">=3.11"

[project.license]
text = "ISC"

[[project.authors]]
name = "thelab"
email = "thelabdev@thelab.co"

[dependency-groups]
dev = [
    "mypy (==1.16.1)",
    "django-cors-headers ==4.7.0",
    "django-stubs (==5.2.1)",
    "djangorestframework-stubs (==3.16.0)",
    "tox>=4.27.0",
    "tox-uv>=1.26.1",
    "ruff (>=0.11.6)",
    "isort (>=6.0.1)",
    "coverage>=7.9.1",
]

[tool.hatch.build.targets.wheel]
packages = ["wagtail_spa_integration"]

[tool.commitizen]
name = "cz_conventional_commits"
annotated_tag = true
gpg_sign = true
tag_format = "v$version"
update_changelog_on_bump = true
changelog_merge_prerelease = true
version_provider = "pep621"
version_scheme = "pep440"
version_files = ["pyproject.toml:version"]
pre_bump_hooks = ["pre-commit run --all-files || true"]
post_bump_hooks = ["git push origin master $CZ_POST_CURRENT_TAG_VERSION"]

[tool.mypy]
python_version = "3.10"
plugins = ["mypy_django_plugin.main"]

# Strict mode, see mypy --help
warn_unused_configs = true
# disallow_subclassing_any = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
show_error_codes = true
# Not turned on by strict
strict_equality = true

[[tool.mypy.overrides]]
module = "*.tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "wagtail.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "django_filters.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "wagtail_headless_preview.models.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sandbox.models"
ignore_errors = true


[tool.django-stubs]
django_settings_module = "sandbox.settings.dev"

[tool.isort]
profile = "black"
from_first = true

[tool.ruff.lint.isort]
from-first = true
